programs/src/requirements.txt
psycopg2-binary==2.9.10
aws-lambda-powertools==3.19.0
aws-xray-sdk==2.5.0
fastjsonschema
jsonpickle
wrapt
boto3

